created route refactor for endpoints and remodeled accounts
